Fix: Extracted components from Home.tsx
